+++ src/library/rendering/openglwrappers.cpp
@@ -22,6 +22,7 @@
 #include "global.h"
 #include "logging.h"
 #include "frame.h"
+#include "screencapture/ScreenCapture.h"
 
 namespace libtas {
 
@@ -168,13 +169,7 @@ GLFUNCSKIPDRAW(glMultiDrawElementsEXT, (GLenum mode, const GLsizei *count, GLenu
 GLFUNCSKIPDRAW(glDrawArraysEXT, (GLenum mode, GLint first, GLsizei count), (mode, first, count))
 #endif
 
-// void glBlitFramebuffer (GLint srcX0, GLint srcY0, GLint srcX1, GLint srcY1, GLint dstX0, GLint dstY0, GLint dstX1, GLint dstY1, GLbitfield mask, GLenum filter)
-// {
-//     LINK_NAMESPACE(glBlitFramebuffer, "GL");
-//     return myglBlitFramebuffer(srcX0, srcY0, srcX1, srcY1, dstX0, dstY0, dstX1, dstY1, mask, filter);
-// }
-
-void myglBlitFramebuffer (GLint srcX0, GLint srcY0, GLint srcX1, GLint srcY1, GLint dstX0, GLint dstY0, GLint dstX1, GLint dstY1, GLbitfield mask, GLenum filter)
+void myglBlitFramebuffer(GLint srcX0, GLint srcY0, GLint srcX1, GLint srcY1, GLint dstX0, GLint dstY0, GLint dstX1, GLint dstY1, GLbitfield mask, GLenum filter)
 {
     DEBUGLOGCALL(LCF_OGL);
     if (!Global::skipping_draw)