+++ src/library/screencapture/ScreenCapture.h
@@ -74,6 +74,11 @@ public:
      * It is equivalent to `copySurfaceToScreen()` in most cases. */
     static void restoreScreenState();
 
+    static void clearScreen();
+
+    /* Return an opaque identifier of the screen texture that was rendered to */
+    static uint32_t screenTexture();
+
 protected:
     
     static ScreenCapture_Impl* impl;